emoji are displayed but without the icons
