+++ src/components/CategoriesNav/categoriesNav.jsx
